assignment/User authentication system/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.41
psycopg2-binary==2.9.10
python-dotenv==1.1.1
pydantic-settings==2.10.1
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose==3.5.0
